+++ b4n-kube/core/mod.rs
@@ -14,9 +14,11 @@ pub const CRDS: &str = "customresourcedefinitions";
 pub use self::kind::{CORE_VERSION, Kind};
 pub use self::namespace::{ALL_NAMESPACES, NAMESPACES, Namespace};
 pub use self::ports::{Port, PortProtocol};
+pub use self::propagation_policy::PropagationPolicy;
 pub use self::resource_ref::{PodRef, ResourceRef, ResourceRefFilter};
 
 mod kind;
 mod namespace;
 mod ports;
+mod propagation_policy;
 mod resource_ref;