+++ b4n-tui/widgets/mod.rs
@@ -1,9 +1,13 @@
+pub use self::actions::{ActionItem, ActionsList, ActionsListBuilder};
 pub use self::footer::Footer;
 pub use self::input::{ErrorHighlightMode, Input};
-pub use self::modal::{Button, CheckBox, ControlsGroup, Dialog};
+pub use self::modal::{Button, CheckBox, Control, ControlsGroup, Dialog, Selector};
+pub use self::select::Select;
 pub use self::validator::{InputValidator, ValidatorKind};
 
+mod actions;
 mod footer;
 mod input;
 mod modal;
+mod select;
 mod validator;