+++ b4n-tui/widgets/modal/mod.rs
@@ -2,8 +2,10 @@ pub use self::button::*;
 pub use self::checkbox::*;
 pub use self::controls_group::*;
 pub use self::dialog::*;
+pub use self::selector::*;
 
 mod button;
 mod checkbox;
 mod controls_group;
 mod dialog;
+mod selector;