+++ b4n/ui/widgets/filters/filter.rs
@@ -1,14 +1,15 @@
 use b4n_common::expr::{ParserError, validate};
 use b4n_config::keys::KeyCommand;
 use b4n_tui::utils::{center_horizontal, get_proportional_width};
+use b4n_tui::widgets::Select;
 use b4n_tui::{MouseEventKind, ResponseEvent, Responsive, TuiEvent, table::Table};
 use crossterm::event::KeyModifiers;
-use ratatui::layout::Rect;
+use ratatui::layout::{Margin, Rect};
 use ratatui::style::Style;
 use ratatui::widgets::{Block, Clear};
 
 use crate::core::{SharedAppData, SharedAppDataExt, SharedBgWorker};
-use crate::ui::widgets::{PatternsList, Select};
+use crate::ui::widgets::PatternsList;
 
 const HISTORY_SIZE: usize = 20;
 
@@ -96,7 +97,7 @@ impl Filter {
         frame.render_widget(Clear, area);
         frame.render_widget(block, area);
 
-        self.patterns.draw(frame, area);
+        self.patterns.draw(frame, area.inner(Margin::new(1, 0)));
     }
 
     /// Validates the filter value as a logical expression.