+++ b4n/ui/widgets/mod.rs
@@ -1,9 +1,7 @@
-pub use self::commands::*;
+pub use self::command_palette::*;
 pub use self::filters::*;
-pub use self::select::*;
 pub use self::side_select::*;
 
-mod commands;
+mod command_palette;
 mod filters;
-mod select;
 mod side_select;